Create mr_robot.cpp
Did this w the help of ChatGPT, and inspiration from my Pwnagotchi. If it works correctly, it’s supposed to be a manual version of one that you can select which network to pwn. If pwned, it will display as such and store the file on the micro SD card. This is the first release and version. H.T.P!!!!